+++ LauncherTests/LauncherTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Ntools;
 using System;
+using System.Diagnostics;
 using System.IO;
 
 namespace Launcher.Tests
@@ -129,6 +130,32 @@ namespace Launcher.Tests
             Assert.IsTrue(result.Output.Contains("fail"));
             Assert.IsTrue(result.Output.Contains("error"));
             Assert.IsTrue(result.Output.Contains("rejected"));
+        }
+
+        [TestMethod()]
+        public void LockVerifyTest()
+        {
+            // Arrange
+            var process = new Process();
+            process.StartInfo.WorkingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.System);
+            process.StartInfo.FileName = $"robocopy.exe";  // must be path rooted
+            process.StartInfo.Arguments = "/?";
+            process.StartInfo.RedirectStandardOutput = true;
+            process.StartInfo.RedirectStandardError = true;
+            process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            process.StartInfo.CreateNoWindow = false;
+
+            var verbose = false;
+            var lockVerify = false;
+
+            // Act
+            var result = process.LockVerifyStart(verbose, lockVerify);
+
+
+            Console.WriteLine($"Output: {result.GetFirstOutput()}");
+            // Assert
+            Assert.AreEqual(16, result.Code);
+            Assert.IsTrue(result.Output.Count > 100);
         }
     }
 }
